refactor(web): split router.tsx into feature modules
router.tsx had grown to 3,170 lines holding the route tree, ~40 page
components, the session contexts, the useResource data hook, Agent
availability projection and every formatting helper. Move each concern to a
module under src/features and leave router.tsx with the route tree and the
three gates, so the following change can turn routes into files.

This is a code move: no behaviour changes and no react-router API changes.

Two source-inspecting tests follow the code they assert on. The Kumo sidebar
contract now reads features/shell/app-shell.tsx, and the PageHeader contract
matches the specifier tail because importers sit at several depths.
