+++ apps/web/src/features/agents/agent-settings/agent-manage-settings.tsx
@@ -1,0 +1,190 @@
+import type { AgentAdminConfig } from "@opentag/shared/browser";
+import { useEffect, useRef, useState } from "react";
+import { useNavigate } from "react-router-dom";
+import { browserApi } from "../../../api.js";
+import {
+  Banner,
+  Button,
+  Dialog,
+  Field,
+  KumoInputControl,
+  SettingsList,
+  SettingsRow,
+  Text,
+} from "../../../ui/design-system.js";
+import type { AgentDetailView } from "../agent-model.js";
+
+export function AgentManageSettings({
+  agent,
+  initialConfig,
+  onAgentChanged,
+}: {
+  agent: AgentDetailView;
+  initialConfig: AgentAdminConfig;
+  onAgentChanged: () => void;
+}) {
+  const navigate = useNavigate();
+  const [config, setConfig] = useState(initialConfig);
+  const [message, setMessage] = useState<string>();
+  const [confirmation, setConfirmation] = useState<"delete" | "pause">();
+  const [confirmationError, setConfirmationError] = useState<string>();
+  const [confirmationText, setConfirmationText] = useState("");
+  const [busy, setBusy] = useState(false);
+  const [restorePauseFocus, setRestorePauseFocus] = useState(false);
+  const pauseButtonRef = useRef<HTMLButtonElement>(null);
+  const deleteButtonRef = useRef<HTMLButtonElement>(null);
+  useEffect(() => {
+    if (confirmation || !restorePauseFocus) return;
+    pauseButtonRef.current?.focus();
+    setRestorePauseFocus(false);
+  }, [confirmation, restorePauseFocus]);
+  async function changeLifecycle(action: "suspend" | "reactivate") {
+    try {
+      setBusy(true);
+      setMessage(undefined);
+      setConfirmationError(undefined);
+      setConfig(
+        action === "suspend" ? await browserApi.suspendAgent(config.id) : await browserApi.reactivateAgent(config.id),
+      );
+      setMessage(action === "suspend" ? "Agent paused." : "Agent reactivated.");
+      setRestorePauseFocus(confirmation === "pause");
+      setConfirmation(undefined);
+      onAgentChanged();
+    } catch (cause) {
+      const error = cause instanceof Error ? cause.message : "Unable to change Agent status";
+      if (confirmation === "pause") setConfirmationError(error);
+      else setMessage(error);
+    } finally {
+      setBusy(false);
+    }
+  }
+  async function deleteAgent() {
+    try {
+      setBusy(true);
+      setMessage(undefined);
+      setConfirmationError(undefined);
+      await browserApi.deleteAgent(config.id);
+      navigate("/agents");
+    } catch (cause) {
+      setConfirmationError(cause instanceof Error ? cause.message : "Unable to delete Agent");
+      setBusy(false);
+    }
+  }
+  function closeConfirmation() {
+    setConfirmation(undefined);
+    setConfirmationError(undefined);
+  }
+  return (
+    <section className="grid gap-4">
+      <header className="grid gap-2">
+        <Text as="h1" size="lg" variant="heading">
+          Manage Agent
+        </Text>
+      </header>
+      <SettingsList>
+        <SettingsRow
+          description={
+            config.status === "active" ? "Stop accepting new requests until reactivated." : "Allow new requests again."
+          }
+          label={config.status === "active" ? "Pause Agent" : "Reactivate Agent"}
+        >
+          <Button
+            ref={pauseButtonRef}
+            variant="secondary"
+            onClick={() => {
+              if (config.status === "active" && agent.activity.state === "working") {
+                setConfirmationError(undefined);
+                setConfirmation("pause");
+                return;
+              }
+              void changeLifecycle(config.status === "active" ? "suspend" : "reactivate");
+            }}
+          >
+            {config.status === "active" ? "Pause" : "Reactivate"}
+          </Button>
+        </SettingsRow>
+        <SettingsRow
+          description={
+            config.status === "active"
+              ? "Pause this Agent before deleting it permanently."
+              : "Permanently remove this Agent. This cannot be undone."
+          }
+          label="Delete Agent"
+        >
+          <Button
+            disabled={config.status === "active"}
+            ref={deleteButtonRef}
+            variant="danger"
+            onClick={() => {
+              setConfirmationText("");
+              setConfirmationError(undefined);
+              setConfirmation("delete");
+            }}
+          >
+            Delete permanently
+          </Button>
+        </SettingsRow>
+      </SettingsList>
+      {message ? <p role="status">{message}</p> : null}
+      {confirmation === "pause" ? (
+        <Dialog
+          busy={busy}
+          description="This Agent is handling a request. Pausing it stops new requests, but the current request may continue until it reaches a safe stopping point."
+          returnFocusRef={pauseButtonRef}
+          title={`Pause ${config.displayName}?`}
+          onClose={closeConfirmation}
+        >
+          {confirmationError ? <Banner variant="error" role="alert" description={confirmationError} /> : null}
+          <div className="flex flex-wrap justify-end gap-3">
+            <Button disabled={busy} variant="ghost" onClick={closeConfirmation}>
+              Keep active
+            </Button>
+            <Button disabled={busy} variant="primary" onClick={() => void changeLifecycle("suspend")}>
+              {busy ? "Pausing…" : "Pause Agent"}
+            </Button>
+          </div>
+        </Dialog>
+      ) : null}
+      {confirmation === "delete" ? (
+        <Dialog
+          busy={busy}
+          description="This permanently removes the Agent and its messaging connection. The Agent cannot be restored."
+          returnFocusRef={deleteButtonRef}
+          title={`Delete ${config.displayName}?`}
+          onClose={closeConfirmation}
+        >
+          <div className="grid gap-4">
+            <Field
+              htmlFor="agent-delete-confirmation"
+              label={
+                <>
+                  Type <strong>{config.displayName}</strong> to confirm
+                </>
+              }
+            >
+              <KumoInputControl
+                autoComplete="off"
+                id="agent-delete-confirmation"
+                value={confirmationText}
+                onChange={(event) => setConfirmationText(event.currentTarget.value)}
+              />
+            </Field>
+            {confirmationError ? <Banner variant="error" role="alert" description={confirmationError} /> : null}
+            <div className="flex flex-wrap justify-end gap-3">
+              <Button disabled={busy} variant="ghost" onClick={closeConfirmation}>
+                Cancel
+              </Button>
+              <Button
+                disabled={busy || confirmationText !== config.displayName}
+                variant="danger"
+                onClick={() => void deleteAgent()}
+              >
+                {busy ? "Deleting…" : "Delete permanently"}
+              </Button>
+            </div>
+          </div>
+        </Dialog>
+      ) : null}
+    </section>
+  );
+}
