+++ apps/web/src/features/agents/agent-settings/sections.ts
@@ -1,0 +1,91 @@
+import type { AgentAdminConfig } from "@opentag/shared/browser";
+import type { IconName } from "../../../ui/design-system.js";
+import type { AgentDetailView } from "../agent-model.js";
+import { messagingConnectionLabel, platformLabel, titleCase } from "../agent-presentation.js";
+
+export type AgentSettingsSection = "instructions" | "execution" | "messaging" | "identity" | "computer" | "manage";
+
+export type AgentSettingsGroup = "work" | "contact" | "details";
+
+export const agentSettingsSections: ReadonlyArray<{
+  key: AgentSettingsSection;
+  label: string;
+  group: AgentSettingsGroup;
+  icon: IconName;
+}> = [
+  {
+    key: "instructions",
+    label: "Instructions",
+    group: "work",
+    icon: "instructions",
+  },
+  {
+    key: "execution",
+    label: "Model & reasoning",
+    group: "work",
+    icon: "model",
+  },
+  {
+    key: "messaging",
+    label: "Messaging",
+    group: "contact",
+    icon: "message",
+  },
+  {
+    key: "identity",
+    label: "Name",
+    group: "details",
+    icon: "user",
+  },
+  {
+    key: "computer",
+    label: "Connected computer",
+    group: "details",
+    icon: "laptop",
+  },
+  {
+    key: "manage",
+    label: "Manage Agent",
+    group: "details",
+    icon: "shield",
+  },
+];
+
+export const agentSettingsGroups = [
+  { key: "work", label: "How it works" },
+  { key: "contact", label: "Where it receives work" },
+  { key: "details", label: "Agent details" },
+] as const;
+
+export function agentSettingsSummary(
+  agent: AgentDetailView,
+  config: AgentAdminConfig,
+  section: AgentSettingsSection,
+): string {
+  if (section === "instructions") {
+    return config.runtimeConfig.instructions.trim() ? "Custom instructions" : "Not configured";
+  }
+  if (section === "execution") {
+    const provider = config.runtimeProvider === "codex" ? "Codex" : "Claude Code";
+    if (!config.runtimeConfig.model && !config.runtimeConfig.reasoningEffort) return `${provider} · Provider defaults`;
+    const model = config.runtimeConfig.model ?? "Default model";
+    const reasoning = config.runtimeConfig.reasoningEffort
+      ? titleCase(config.runtimeConfig.reasoningEffort)
+      : "Default reasoning";
+    return `${provider} · ${model} · ${reasoning}`;
+  }
+  if (section === "messaging") {
+    if (agent.messaging.kind === "unconfirmed") return "Messaging status is temporarily unavailable";
+    const binding = agent.messaging.value;
+    if (!binding) return "No messaging channel connected";
+    const status = messagingConnectionLabel(binding);
+    return `${titleCase(binding.provider)} · @${agent.name} · ${status}`;
+  }
+  if (section === "identity") return config.displayName;
+  if (section === "computer") {
+    const state = agent.availability.dependencies.computer.state;
+    const status = state === "ready" ? "Online" : state === "action_required" ? "Offline" : "Unconfirmed";
+    return `${agent.computer.displayName} · ${platformLabel(agent.computer.platform)} · ${status}`;
+  }
+  return config.status === "active" ? "Active" : "Paused";
+}
